Breaking: don't skip downloads in standalone mode
I.e. when running `npm install` in the working directory of a
project that uses prebuild-install, you must now run a more
explicit `npm install --build-from-source`.

This is necessary to support npm 7, where we do not have enough
information to automatically determine it and would erroneously
skip downloads on normal installs.